refact(engine): engine refactored
