+++ src/engine/core/assets.ts
